pyproject.toml
[project]
name = "azure-kinect-3d"
version = "0.1.0"
description = "RGBD camera + SAM3 segmentation + 3D reconstruction pipeline"
requires-python = ">=3.12"
dependencies = [
    "numpy<2",
    "trimesh",
    "scipy",
    "torch>=2.0",
    "torchvision>=0.15",
    "Pillow",
    "sam3",
    "einops>=0.8.2",
    "pycocotools",
    "psutil>=7.2.2",
    "pyrealsense2>=2.57.7.10387",
    "hydra-core",
    "omegaconf",
    "loguru",
]

[tool.uv.sources]
sam3 = { path = "third_party/sam3", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_kinect_3d"]

[tool.uv]
index-strategy = "unsafe-best-match"
environments = ["sys_platform == 'linux'"]
# sam3 pins ftfy==6.1.1, we need >=6.1.1 for flexibility
override-dependencies = [
    "ftfy>=6.1.1",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources.torch]
index = "pytorch"

[tool.uv.sources.torchvision]
index = "pytorch"
